+++ CHANGELOG.md
@@ -13,11 +13,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `WebRtcMediaConnectionFactory.createLocalVideoTrack` that accepts `VideoCapturer`
 - `ParticipantStep.message` that allows sending direct messages
 - `MessageReceivedEvent.direct` that indicates whether it's direct message
+- `Messenger` that provides facilities to send and receive `Message`s; note that unlike previous
+  implementation, `MessageListener` will not be invoked when you send a message
 
 ### Changed
 
 - WebRTC to M110
 - **BREAKING**: `MessageRequest` no longer has a default `type` set
+- Deprecated `Conference.message(String)`
 
 ### Fixed